fix: hide AI assistant on EU Access Only clusters across all surfaces
